STEP 117 = DONE

+++ student_kalesnik_ivan_v2/src/test/resources/schema.sql
@@ -84,9 +84,9 @@ CREATE UNIQUE INDEX ix_selected_risks_agreement_id_risk_ic
     ON selected_risks(agreement_id, risk_ic);
 
 CREATE TABLE agreement_persons (
-                                   id BIGINT NOT NULL AUTO_INCREMENT,
-                                   agreement_id BIGINT NOT NULL,
-                                   person_id BIGINT NOT NULL,
+                                   id NUMERIC NOT NULL AUTO_INCREMENT,
+                                   agreement_id NUMERIC NOT NULL,
+                                   person_id NUMERIC NOT NULL,
                                    medical_risk_limit_level VARCHAR(100) NOT NULL,
                                    PRIMARY KEY (id),
                                    foreign key (agreement_id) references agreements(id),
